requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
boto3==1.26.40
botocore==1.29.40
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.85.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.0
Pillow==9.3.0
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uuid==1.30
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.3
